speed up module bootstrapping  … (#26)
* update vendor

* speed up module bootstrapping

Before we were doing a fetch on the module every time. On fast broadband this
is like 3-4 sec. But on tethering it is 100sec+.

This diff adds caching for modules and cached speed is negligable.

I couldn't directly use the terraform because it isn't well encapsulated, so
this diff means we don't support terraform registry modules and we only invaliate
the cache on new versions of fogg. This is ok because we only support the 1
hardcoded module for now.

* improve error handling