full intgration test
This is a bit of an experiment, but I think it might be useful to have full-tool integration tests
that exercise the whole (almost) process of running commands and compares the outputs. Might be
useful to catch regressions before we get to the point of rolling out on users' repos.

Not sure if this will be fun to maintain+grow or not, but I started with a very minimal test (the
equivalent of `fogg init`) to see how it goes.