[feature] optional tflint support  (#134)
 [feature] optional tflint support Consider this experiement for now, but if you set "tflint": {"enabled": true} you will
get a `make lint-tflint` which runs [tflint](https://github.com/wata727/tflint).

We can start opting into this and if it works well we can make it the default and/or make the config
a bit simpler (add the plugin automatically if it is turned on).

This diff also switches plant.Print to YAML and turns on `terraform validate` in our make files.

Fixes #125 
Fixes #132