[feature] check for version mis-match on apply (#144)
[feature] check for version mis-match on apply
This will check to see if the version in .fogg-version matches the current tool version. If --upgrade/-u is not supplied, apply will fail.

Currently the UX for this error is bad. After the PR I will do a pass at better error UX before we do a release.

Fixes #51