[feature] [breaking] fogg setup (#204)
In order to support running outside of docker containers, we have a new command fogg setup which will download and install custom plugins (and anything else we need in the future).

Note that this means "setup the current working copy of this repo" not "setup this repo".

This is a biggish change and includes a few yak-shaves–

new command setup led to some refactoring of how we bootstrap the commands. There is probably room to DRY up the other commands, but will leave that to a follow-up.
The apply command no longer does plugins so we no longer need the --no-apply-plugins flag.
We now template custom plugins stuff to inject the OS and Arch into them. Example of how to use this is here