golden file tests (#242)
As a way to tackle #87, #88 , etc, this introduces some "golden file tests" (Background).

The basic idea is that we have fogg.json files in testdata and can run apply over them and check them in (make update-golden-files). Then when we run the tests we apply to a temp directory and compare the results.

The code for comparing has some weaknesses, but is a starting point for making sure we don't introduce (unknown) changes to the outputted files.