[breaking] make aws provider optional
This change makes it optional to specify an aws provider. Only planning on supporting this for config
v2 but it might work in v1 for some cases.

The only breaking change is that we now require an aws account id for the global component, where before
we would silently ignore the fact that it was missing.