[breaking] make aws provider optional (#247)
[breaking] make aws provider optional### Summary
This change makes it optional to specify an aws provider. Only planning on supporting this for config
v2 but it might accidentally work in v1.

The only breaking change is that we now require an aws account id for the global component, where before we would silently ignore the fact that it was missing. I consider this a bug / oversight so fixing it here.

### Test Plan

This diff adds a number of tests, especially golden file tests. None of the existing golden files had to be changed.

### References

*