[fix] Create directories in aws config path file if they doesn't exis… (#397)
[fix] Create directories in aws config path file if they doesn't exis……t already

### Summary
We are using `os.OpenFile` to create the awsconfig file. This will fail if any directories in the path doesn't exist as it is happening in my case. Added code to create any non existing directories in the path.

### Test Plan
Tested the code change locally.

### References
(Optional) Additional links to provide more context.