[docs] update readme on releasing, etc (#432)
[docs] update readme on releasing, etcThis adds a section to README.md that documents how to release this
package.

Additionally we update the copyright and remove windows from the builds.
Happy to add back windows if we can find a way to test it automatically.

### Summary
A description of changes or issues addressed by this PR. Provide links to relevant PRs if any.

### Test Plan
Say unittests, or list out steps to verify changes.

### References
(Optional) Additional links to provide more context.