[feature] automatically prompt for heroku login (#452)
[feature] automatically prompt for heroku loginFor users of heroku SSO, the login session times out fairly quickly.
This means that almost every time we use terraform, whether for a plan
or an apply, we see the "login to heroku" error and need to re-run the
command.

This changes the fogg defaults to prompt for login, inline, so that it's
possible to click the SSO link, authenticate, and continue with the
original `make` command uninterrupted.

For non SSO users, this probably will be a no-op, except on their first
run, if they are also not logged into heroku at that point, which seems
really unlikely.

Test Plan:
make
fogg version to confirm latest updates
switch to fogg managed repo
fogg apply
`git grep "heroku auth"` -> observe new command in templates
`make plan` and see that auth is prompted.