[feature] support configuring aws providers with roles (#469)
[feature] support configuring aws providers with rolesRather than specifying a profile name, you can now specify a role name.
That role name combined with the account id is used to create the arn
for a role to assume.

Previously we only supported using profile names for configuring aws
providers. This allows late-binding of roles wherein you can use the
exact same code but different aws iam principals in different
circumstances. Example - a read-only role for running a plan check in
CI.

For using Terraform Enterprise we want to switch to specifying a role
rather than trying to thread through an aws cli config file.

Fixes #468

### Test Plan
* [x] CI
* [x] manually tested a few internal repos

### References