[fix] simplify and optimize .terraformignore
Make the configuation simplier and 10x faster at building slugs.

This is faster because the process seems to be CPU bound when you have a
large number of patterns in your .terraformignore.

And slug sizes are smaller because we now exclude modules directorties.

Tested on shared-infra and runtime when from 6m -> 12s and slug size
from 75MB to 5MB.