[fix] simplify and optimize .terraformignore (#483)
Make the configuration simpler and 10x faster at building slugs.

This is faster because the process seems to be CPU bound when you have a
large number of patterns in your .terraformignore.

And slug sizes are smaller because we now exclude modules directorties.

### Summary
A description of changes or issues addressed by this PR. Provide links to relevant PRs if any.

### Test Plan

Tested on shared-infra and runtime when from 6m -> 12s and slug size
from 75MB to 5MB.


### References
* library that builds the slugs - https://github.com/hashicorp/go-slug