[feature] Allow overriding name of module when using module_source (#484)
### Summary
When using module_source in a component, allow providing the name of the module that is autogenerated by fogg, by providing a "module_name" field. This is useful if you want to use a different module name than the name matching the subdirectory. This can be desired when you have similar modules with different names but mostly compatible contents, and you don't want to force a user to do state mv migrations between the modules.

If module_name is not provided or is empty string, falls back to current default behavior of using the last directory in the module's subdirectory path as the module name

### Test Plan
Unit tests added