[feature] auto remote state (experimental) (#498)
Read a directory of terraform code, find the remote state references and update depends_on in fogg.yml with them.

This code is *very* rough but working for the happy paths. I till do a full parse of code in a terraform directory (in a fogg-managed repo) and extract all the references to `data.terraform_remote_state.foo`. Those will then be used to update depends_on configurations in fogg.yml. It is expected that the user would then run `fogg apply` to update.

Example run, using [dyff](https://github.com/homeport/dyff) to visualize the yaml diff -

```
$ $GOPATH/bin/fogg exp auto-remote-state --path terraform/accounts/czi-id
$ git diff

accounts.czi-id
  + one map entry added:
    depends_on:
    │ accounts:
    │ - okta-czi-duo
```

### Test Plan

I've tested this manually on a few repos and components, working to do some more.

### References