chore: Add release please version manifest
The simplest way to tell release-please the current version for a package when switching to manifest releases is to manually add an entry into .release-please-manifest.json. This change should be made directly on the default/configured branch or on a separate, user-created branch/PR which is then merged into the default/configured branch.

Ref:

- [Initial version](https://github.com/googleapis/release-please/blob/main/docs/manifest-releaser.md#initial-version)